Crate linera_service

This module provides the executables needed to operate a Linera service, including a placeholder wallet acting as a GraphQL service for user interfaces.

Re-exports

pub use wallet::Wallet;

Modules

cli

Helper module for the Linera CLI binary.

cli_wrappers

Helper module to call the binaries of linera-service with appropriate command-line arguments.

config

controller

node_service

project

storage

task_processor

Task processor for executing off-chain operators on behalf of on-chain applications.

tracing

This module provides unified handling for tracing subscribers within Linera binaries.

util

wallet

Macros

test_name