linera_service::cli::command

Enum ClientCommand

pub enum ClientCommand {

    Transfer {
        sender: Account,
        recipient: Account,
        amount: Amount,
    },
    OpenChain {
        chain_id: Option<ChainId>,
        owner: Option<AccountOwner>,
        balance: Amount,
        super_owner: bool,
    },
    OpenMultiOwnerChain {
        chain_id: Option<ChainId>,
        ownership_config: ChainOwnershipConfig,
        application_permissions_config: ApplicationPermissionsConfig,
        balance: Amount,
    },
    ShowOwnership {
        chain_id: Option<ChainId>,
    },
    ChangeOwnership {
        chain_id: Option<ChainId>,
        ownership_config: ChainOwnershipConfig,
    },
    SetPreferredOwner {
        chain_id: Option<ChainId>,
        owner: AccountOwner,
    },
    ChangeApplicationPermissions {
        chain_id: Option<ChainId>,
        application_permissions_config: ApplicationPermissionsConfig,
    },
    CloseChain {
        chain_id: ChainId,
    },
    ShowNetworkDescription,
    LocalBalance {
        account: Option<Account>,
    },
    QueryBalance {
        account: Option<Account>,
    },
    SyncBalance {
        account: Option<Account>,
    },
    Sync {
        chain_id: Option<ChainId>,
    },
    ProcessInbox {
        chain_id: Option<ChainId>,
    },
    QueryShardInfo {
        chain_id: ChainId,
    },
    RevokeEpochs {
        epoch: Epoch,
    },
    ResourceControlPolicy {

        wasm_fuel_unit: Option<Amount>,
        evm_fuel_unit: Option<Amount>,
        read_operation: Option<Amount>,
        write_operation: Option<Amount>,
        byte_runtime: Option<Amount>,
        byte_read: Option<Amount>,
        byte_written: Option<Amount>,
        blob_read: Option<Amount>,
        blob_published: Option<Amount>,
        blob_byte_read: Option<Amount>,
        blob_byte_published: Option<Amount>,
        byte_stored: Option<Amount>,
        operation: Option<Amount>,
        operation_byte: Option<Amount>,
        message: Option<Amount>,
        message_byte: Option<Amount>,
        service_as_oracle_query: Option<Amount>,
        http_request: Option<Amount>,
        maximum_wasm_fuel_per_block: Option<u64>,
        maximum_evm_fuel_per_block: Option<u64>,
        maximum_service_oracle_execution_ms: Option<u64>,
        maximum_block_size: Option<u64>,
        maximum_blob_size: Option<u64>,
        maximum_published_blobs: Option<u64>,
        maximum_bytecode_size: Option<u64>,
        maximum_block_proposal_size: Option<u64>,
        maximum_bytes_read_per_block: Option<u64>,
        maximum_bytes_written_per_block: Option<u64>,
        maximum_oracle_response_bytes: Option<u64>,
        maximum_http_response_bytes: Option<u64>,
        http_request_timeout_ms: Option<u64>,
        http_request_allow_list: Option<Vec<String>>,
    
},
    Benchmark(BenchmarkCommand),
    CreateGenesisConfig {

        committee_config_path: PathBuf,
        genesis_config_path: PathBuf,
        initial_funding: Amount,
        start_timestamp: Option<DateTime<Utc>>,
        num_other_initial_chains: u32,
        policy_config: ResourceControlPolicyConfig,
        wasm_fuel_unit_price: Option<Amount>,
        evm_fuel_unit_price: Option<Amount>,
        read_operation_price: Option<Amount>,
        write_operation_price: Option<Amount>,
        byte_runtime_price: Option<Amount>,
        byte_read_price: Option<Amount>,
        byte_written_price: Option<Amount>,
        blob_read_price: Option<Amount>,
        blob_published_price: Option<Amount>,
        blob_byte_read_price: Option<Amount>,
        blob_byte_published_price: Option<Amount>,
        byte_stored_price: Option<Amount>,
        operation_price: Option<Amount>,
        operation_byte_price: Option<Amount>,
        message_price: Option<Amount>,
        message_byte_price: Option<Amount>,
        service_as_oracle_query_price: Option<Amount>,
        http_request_price: Option<Amount>,
        maximum_wasm_fuel_per_block: Option<u64>,
        maximum_evm_fuel_per_block: Option<u64>,
        maximum_service_oracle_execution_ms: Option<u64>,
        maximum_block_size: Option<u64>,
        maximum_bytecode_size: Option<u64>,
        maximum_blob_size: Option<u64>,
        maximum_published_blobs: Option<u64>,
        maximum_block_proposal_size: Option<u64>,
        maximum_bytes_read_per_block: Option<u64>,
        maximum_bytes_written_per_block: Option<u64>,
        maximum_oracle_response_bytes: Option<u64>,
        maximum_http_response_bytes: Option<u64>,
        http_request_timeout_ms: Option<u64>,
        http_request_allow_list: Option<Vec<String>>,
        testing_prng_seed: Option<u64>,
        network_name: Option<String>,
    
},
    Watch {
        chain_id: Option<ChainId>,
        raw: bool,
    },
    Service {
        config: ChainListenerConfig,
        port: NonZeroU16,
        metrics_port: NonZeroU16,
        operator_application_ids: Vec<ApplicationId>,
        controller_application_id: Option<ApplicationId>,
        operators: Vec<(String, PathBuf)>,
    },
    Faucet {
        chain_id: Option<ChainId>,
        port: u16,
        metrics_port: u16,
        amount: Amount,
        limit_rate_until: Option<DateTime<Utc>>,
        config: ChainListenerConfig,
        storage_path: PathBuf,
        max_batch_size: usize,
    },
    PublishModule {
        contract: PathBuf,
        service: PathBuf,
        vm_runtime: VmRuntime,
        publisher: Option<ChainId>,
    },
    ListEventsFromIndex {
        chain_id: Option<ChainId>,
        stream_id: StreamId,
        start_index: u32,
    },
    PublishDataBlob {
        blob_path: PathBuf,
        publisher: Option<ChainId>,
    },
    ReadDataBlob {
        hash: CryptoHash,
        reader: Option<ChainId>,
    },
    CreateApplication {
        module_id: ModuleId,
        creator: Option<ChainId>,
        json_parameters: Option<String>,
        json_parameters_path: Option<PathBuf>,
        json_argument: Option<String>,
        json_argument_path: Option<PathBuf>,
        required_application_ids: Option<Vec<ApplicationId>>,
    },
    PublishAndCreate {
        contract: PathBuf,
        service: PathBuf,
        vm_runtime: VmRuntime,
        publisher: Option<ChainId>,
        json_parameters: Option<String>,
        json_parameters_path: Option<PathBuf>,
        json_argument: Option<String>,
        json_argument_path: Option<PathBuf>,
        required_application_ids: Option<Vec<ApplicationId>>,
    },
    Keygen,
    Assign {
        owner: AccountOwner,
        chain_id: ChainId,
    },
    RetryPendingBlock {
        chain_id: Option<ChainId>,
    },
    Wallet(WalletCommand),
    Chain(ChainCommand),
    Project(ProjectCommand),
    Net(NetCommand),
    Validator(Command),
    Storage(DatabaseToolCommand),
    HelpMarkdown,
    ExtractScriptFromMarkdown {
        path: PathBuf,
        pause_after_linera_service: Duration,
        pause_after_gql_mutations: Duration,
    },
    Completion {
        shell: Shell,
    },

}

Variants

Transfer

Transfer funds

Fields

sender: Account

Sending chain ID (must be one of our chains)

recipient: Account

Recipient account

amount: Amount

Amount to transfer

OpenChain

Open (i.e. activate) a new chain deriving the UID from an existing one.

Fields

chain_id: Option<ChainId>

Chain ID (must be one of our chains).

owner: Option<AccountOwner>

The new owner (otherwise create a key pair and remember it)

balance: Amount

The initial balance of the new chain. This is subtracted from the parent chains balance.

super_owner: bool

Whether to create a super owner for the new chain.

OpenMultiOwnerChain

Open (i.e. activate) a new multi-owner chain deriving the UID from an existing one.

Fields

chain_id: Option<ChainId>

Chain ID (must be one of our chains).

ownership_config: ChainOwnershipConfig

application_permissions_config: ApplicationPermissionsConfig

balance: Amount

The initial balance of the new chain. This is subtracted from the parent chains balance.

ShowOwnership

Display who owns the chain, and how the owners work together proposing blocks.

Fields

chain_id: Option<ChainId>

The ID of the chain whose owners will be changed.

ChangeOwnership

Change who owns the chain, and how the owners work together proposing blocks.

Specify the complete set of new owners, by public key. Existing owners that are not included will be removed.

Fields

chain_id: Option<ChainId>

The ID of the chain whose owners will be changed.

ownership_config: ChainOwnershipConfig

SetPreferredOwner

Change the preferred owner of a chain.

Fields

chain_id: Option<ChainId>

The ID of the chain whose preferred owner will be changed.

owner: AccountOwner

The new preferred owner.

ChangeApplicationPermissions

Changes the application permissions configuration.

Fields

chain_id: Option<ChainId>

The ID of the chain to which the new permissions will be applied.

application_permissions_config: ApplicationPermissionsConfig

CloseChain

Close an existing chain.

A closed chain cannot execute operations or accept messages anymore. It can still reject incoming messages, so they bounce back to the sender.

Fields

chain_id: ChainId

Chain ID (must be one of our chains)

ShowNetworkDescription

Print out the network description.

LocalBalance

Read the current native-token balance of the given account directly from the local state.

NOTE: The local balance does not reflect messages that are waiting to be picked in the local inbox, or that have not been synchronized from validators yet. Use linera sync then either linera query-balance or linera process-inbox && linera local-balance for a consolidated balance.

Fields

account: Option<Account>

The account to read, written as CHAIN-ID:OWNER or simply CHAIN-ID for the chain balance. By default, we read the chain balance of the default chain in the wallet.

QueryBalance

Simulate the execution of one block made of pending messages from the local inbox, then read the native-token balance of the account from the local state.

NOTE: The balance does not reflect messages that have not been synchronized from validators yet. Call linera sync first to do so.

Fields

account: Option<Account>

The account to query, written as CHAIN-ID:OWNER or simply CHAIN-ID for the chain balance. By default, we read the chain balance of the default chain in the wallet.

SyncBalance

(DEPRECATED) Synchronize the local state of the chain with a quorum validators, then query the local balance.

This command is deprecated. Use linera sync && linera query-balance instead.

Fields

account: Option<Account>

The account to query, written as CHAIN-ID:OWNER or simply CHAIN-ID for the chain balance. By default, we read the chain balance of the default chain in the wallet.

Sync

Synchronize the local state of the chain with a quorum validators.

Fields

chain_id: Option<ChainId>

The chain to synchronize with validators. If omitted, synchronizes the default chain of the wallet.

ProcessInbox

Process all pending incoming messages from the inbox of the given chain by creating as many blocks as needed to execute all (non-failing) messages. Failing messages will be marked as rejected and may bounce to their sender depending on their configuration.

Fields

chain_id: Option<ChainId>

The chain to process. If omitted, uses the default chain of the wallet.

QueryShardInfo

Query validators for shard information about a specific chain.

Fields

chain_id: ChainId

The chain to query shard information for.

RevokeEpochs

Deprecates all committees up to and including the specified one.

Fields

epoch: Epoch

ResourceControlPolicy

View or update the resource control policy

Fields

wasm_fuel_unit: Option<Amount>

Set the price per unit of Wasm fuel.

evm_fuel_unit: Option<Amount>

Set the price per unit of EVM fuel.

read_operation: Option<Amount>

Set the price per read operation.

write_operation: Option<Amount>

Set the price per write operation.

byte_runtime: Option<Amount>

Set the price per byte read from runtime.

byte_read: Option<Amount>

Set the price per byte read.

byte_written: Option<Amount>

Set the price per byte written.

blob_read: Option<Amount>

Set the base price to read a blob.

blob_published: Option<Amount>

Set the base price to publish a blob.

blob_byte_read: Option<Amount>

Set the price to read a blob, per byte.

blob_byte_published: Option<Amount>

The price to publish a blob, per byte.

byte_stored: Option<Amount>

Set the price per byte stored.

operation: Option<Amount>

Set the base price of sending an operation from a block..

operation_byte: Option<Amount>

Set the additional price for each byte in the argument of a user operation.

message: Option<Amount>

Set the base price of sending a message from a block..

message_byte: Option<Amount>

Set the additional price for each byte in the argument of a user message.

service_as_oracle_query: Option<Amount>

Set the price per query to a service as an oracle.

http_request: Option<Amount>

Set the price for performing an HTTP request.

maximum_wasm_fuel_per_block: Option<u64>

Set the maximum amount of Wasm fuel per block.

maximum_evm_fuel_per_block: Option<u64>

Set the maximum amount of EVM fuel per block.

maximum_service_oracle_execution_ms: Option<u64>

Set the maximum time in milliseconds that a block can spend executing services as oracles.

maximum_block_size: Option<u64>

Set the maximum size of a block, in bytes.

maximum_blob_size: Option<u64>

Set the maximum size of data blobs, compressed bytecode and other binary blobs, in bytes.

maximum_published_blobs: Option<u64>

Set the maximum number of published blobs per block.

maximum_bytecode_size: Option<u64>

Set the maximum size of decompressed contract or service bytecode, in bytes.

maximum_block_proposal_size: Option<u64>

Set the maximum size of a block proposal, in bytes.

maximum_bytes_read_per_block: Option<u64>

Set the maximum read data per block.

maximum_bytes_written_per_block: Option<u64>

Set the maximum write data per block.

maximum_oracle_response_bytes: Option<u64>

Set the maximum size of oracle responses.

maximum_http_response_bytes: Option<u64>

Set the maximum size in bytes of a received HTTP response.

http_request_timeout_ms: Option<u64>

Set the maximum amount of time allowed to wait for an HTTP response.

http_request_allow_list: Option<Vec<String>>

Set the list of hosts that contracts and services can send HTTP requests to.

Benchmark(BenchmarkCommand)

Run benchmarks to test network performance.

CreateGenesisConfig

Create genesis configuration for a Linera deployment. Create initial user chains and print information to be used for initialization of validator setup. This will also create an initial wallet for the owner of the initial root chains.

Fields

committee_config_path: PathBuf

Sets the file describing the public configurations of all validators

genesis_config_path: PathBuf

The output config path to be consumed by the server

initial_funding: Amount

Known initial balance of the chain

start_timestamp: Option<DateTime<Utc>>

The start timestamp: no blocks can be created before this time.

num_other_initial_chains: u32

Number of initial (aka root) chains to create in addition to the admin chain.

policy_config: ResourceControlPolicyConfig

Configure the resource control policy (notably fees) according to pre-defined settings.

wasm_fuel_unit_price: Option<Amount>

Set the price per unit of Wasm fuel. (This will overwrite value from --policy-config)

evm_fuel_unit_price: Option<Amount>

Set the price per unit of EVM fuel. (This will overwrite value from --policy-config)

read_operation_price: Option<Amount>

Set the price per read operation. (This will overwrite value from --policy-config)

write_operation_price: Option<Amount>

Set the price per write operation. (This will overwrite value from --policy-config)

byte_runtime_price: Option<Amount>

Set the price per byte read from runtime. (This will overwrite value from --policy-config)

byte_read_price: Option<Amount>

Set the price per byte read. (This will overwrite value from --policy-config)

byte_written_price: Option<Amount>

Set the price per byte written. (This will overwrite value from --policy-config)

blob_read_price: Option<Amount>

Set the base price to read a blob. (This will overwrite value from --policy-config)

blob_published_price: Option<Amount>

Set the base price to publish a blob. (This will overwrite value from --policy-config)

blob_byte_read_price: Option<Amount>

Set the price to read a blob, per byte. (This will overwrite value from --policy-config)

blob_byte_published_price: Option<Amount>

Set the price to publish a blob, per byte. (This will overwrite value from --policy-config)

byte_stored_price: Option<Amount>

Set the price per byte stored. (This will overwrite value from --policy-config)

operation_price: Option<Amount>

Set the base price of sending an operation from a block.. (This will overwrite value from --policy-config)

operation_byte_price: Option<Amount>

Set the additional price for each byte in the argument of a user operation. (This will overwrite value from --policy-config)

message_price: Option<Amount>

Set the base price of sending a message from a block.. (This will overwrite value from --policy-config)

message_byte_price: Option<Amount>

Set the additional price for each byte in the argument of a user message. (This will overwrite value from --policy-config)

service_as_oracle_query_price: Option<Amount>

Set the price per query to a service as an oracle.

http_request_price: Option<Amount>

Set the price for performing an HTTP request.

maximum_wasm_fuel_per_block: Option<u64>

Set the maximum amount of Wasm fuel per block. (This will overwrite value from --policy-config)

maximum_evm_fuel_per_block: Option<u64>

Set the maximum amount of EVM fuel per block. (This will overwrite value from --policy-config)

maximum_service_oracle_execution_ms: Option<u64>

Set the maximum time in milliseconds that a block can spend executing services as oracles.

maximum_block_size: Option<u64>

Set the maximum size of a block. (This will overwrite value from --policy-config)

maximum_bytecode_size: Option<u64>

Set the maximum size of decompressed contract or service bytecode, in bytes. (This will overwrite value from --policy-config)

maximum_blob_size: Option<u64>

Set the maximum size of data blobs, compressed bytecode and other binary blobs, in bytes. (This will overwrite value from --policy-config)

maximum_published_blobs: Option<u64>

Set the maximum number of published blobs per block. (This will overwrite value from --policy-config)

maximum_block_proposal_size: Option<u64>

Set the maximum size of a block proposal, in bytes. (This will overwrite value from --policy-config)

maximum_bytes_read_per_block: Option<u64>

Set the maximum read data per block. (This will overwrite value from --policy-config)

maximum_bytes_written_per_block: Option<u64>

Set the maximum write data per block. (This will overwrite value from --policy-config)

maximum_oracle_response_bytes: Option<u64>

Set the maximum size of oracle responses. (This will overwrite value from --policy-config)

maximum_http_response_bytes: Option<u64>

Set the maximum size in bytes of a received HTTP response.

http_request_timeout_ms: Option<u64>

Set the maximum amount of time allowed to wait for an HTTP response.

http_request_allow_list: Option<Vec<String>>

Set the list of hosts that contracts and services can send HTTP requests to.

testing_prng_seed: Option<u64>

Force this wallet to generate keys using a PRNG and a given seed. USE FOR TESTING ONLY.

network_name: Option<String>

A unique name to identify this network.

Watch

Watch the network for notifications.

Fields

chain_id: Option<ChainId>

The chain ID to watch.

raw: bool

Show all notifications from all validators.

Service

Run a GraphQL service to explore and extend the chains of the wallet.

Fields

config: ChainListenerConfig

port: NonZeroU16

The port on which to run the server

metrics_port: NonZeroU16

The port to expose metrics on.

operator_application_ids: Vec<ApplicationId>

Application IDs of operator applications to watch. When specified, a task processor is started alongside the node service.

controller_application_id: Option<ApplicationId>

A controller to execute a dynamic set of applications running on a dynamic set of chains.

operators: Vec<(String, PathBuf)>

Supported operators and their binary paths. Format: name=path or just name (uses name as path). Example: --operators my-operator=/path/to/binary

Faucet

Run a GraphQL service that exposes a faucet where users can claim tokens. This gives away the chains tokens, and is mainly intended for testing.

Fields

chain_id: Option<ChainId>

The chain that gives away its tokens.

port: u16

The port on which to run the server

metrics_port: u16

The port for prometheus to scrape.

amount: Amount

The number of tokens to send to each new chain.

limit_rate_until: Option<DateTime<Utc>>

The end timestamp: The faucet will rate-limit the token supply so it runs out of money no earlier than this.

config: ChainListenerConfig

Configuration for the faucet chain listener.

storage_path: PathBuf

Path to the persistent storage file for faucet mappings.

max_batch_size: usize

Maximum number of operations to include in a single block (default: 100).

PublishModule

Publish module.

Fields

contract: PathBuf

Path to the Wasm file for the application contract bytecode.

service: PathBuf

Path to the Wasm file for the application service bytecode.

vm_runtime: VmRuntime

The virtual machine runtime to use.

publisher: Option<ChainId>

An optional chain ID to publish the module. The default chain of the wallet is used otherwise.

ListEventsFromIndex

Print events from a specific chain and stream from a specified index.

Fields

chain_id: Option<ChainId>

The chain to query. If omitted, query the default chain of the wallet.

stream_id: StreamId

The stream being considered.

start_index: u32

Index of the message to start with

PublishDataBlob

Publish a data blob of binary data.

Fields

blob_path: PathBuf

Path to data blob file to be published.

publisher: Option<ChainId>

An optional chain ID to publish the blob. The default chain of the wallet is used otherwise.

ReadDataBlob

Verify that a data blob is readable.

Fields

hash: CryptoHash

The hash of the content.

reader: Option<ChainId>

An optional chain ID to verify the blob. The default chain of the wallet is used otherwise.

CreateApplication

Create an application.

Fields

module_id: ModuleId

The module ID of the application to create.

creator: Option<ChainId>

An optional chain ID to host the application. The default chain of the wallet is used otherwise.

json_parameters: Option<String>

The shared parameters as JSON string.

json_parameters_path: Option<PathBuf>

Path to a JSON file containing the shared parameters.

json_argument: Option<String>

The instantiation argument as a JSON string.

json_argument_path: Option<PathBuf>

Path to a JSON file containing the instantiation argument.

required_application_ids: Option<Vec<ApplicationId>>

The list of required dependencies of application, if any.

PublishAndCreate

Create an application, and publish the required module.

Fields

contract: PathBuf

Path to the Wasm file for the application contract bytecode.

service: PathBuf

Path to the Wasm file for the application service bytecode.

vm_runtime: VmRuntime

The virtual machine runtime to use.

publisher: Option<ChainId>

An optional chain ID to publish the module. The default chain of the wallet is used otherwise.

json_parameters: Option<String>

The shared parameters as JSON string.

json_parameters_path: Option<PathBuf>

Path to a JSON file containing the shared parameters.

json_argument: Option<String>

The instantiation argument as a JSON string.

json_argument_path: Option<PathBuf>

Path to a JSON file containing the instantiation argument.

required_application_ids: Option<Vec<ApplicationId>>

The list of required dependencies of application, if any.

Keygen

Create an unassigned key pair.

Assign

Link the owner to the chain. Expects that the caller has a private key corresponding to the public_key, otherwise block proposals will fail when signing with it.

Fields

owner: AccountOwner

The owner to assign.

chain_id: ChainId

The ID of the chain.

RetryPendingBlock

Retry a block we unsuccessfully tried to propose earlier.

As long as a block is pending most other commands will fail, since it is unsafe to propose multiple blocks at the same height.

Fields

chain_id: Option<ChainId>

The chain with the pending block. If not specified, the wallets default chain is used.

Wallet(WalletCommand)

Show the contents of the wallet.

Chain(ChainCommand)

Show the information about a chain.

Project(ProjectCommand)

Manage Linera projects.

Net(NetCommand)

Manage a local Linera Network.

Validator(Command)

Manage validators in the committee.

Storage(DatabaseToolCommand)

Operation on the storage.

HelpMarkdown

Print CLI help in Markdown format, and exit.

ExtractScriptFromMarkdown

Extract a Bash and GraphQL script embedded in a markdown file and print it on stdout.

Fields

path: PathBuf

The source file

pause_after_linera_service: Duration

Insert a pause of N seconds after calls to linera service.

pause_after_gql_mutations: Duration

Insert a pause of N seconds after GraphQL queries.

Completion

Generate shell completion scripts

Fields

shell: Shell

The shell to generate completions for

Implementations

impl ClientCommand

pub fn log_file_name(&self) -> Cow<'static, str>

Returns the log file name to use based on the ClientCommand that will run.