Module linera_service::cli_wrappers::local_net

How to run Linera validators locally as native processes.

Structs

LocalNet

A set of Linera validators running locally as native processes.

LocalNetConfig

The information needed to start a LocalNet.

Enums

Database

Description of the database engine to use inside a local Linera network.

PathProvider

Path used for the run can come from a path whose lifetime is controlled by an external user or as a temporary directory

ProcessInbox

StorageConfigBuilder

Functions

get_node_port

Provides a port for the node_service. Increment the port numbers.