linera_service::tracing::chrome

Type Alias ChromeTraceGuard

pub type ChromeTraceGuard = FlushGuard;

Guard that flushes Chrome trace file when dropped.

Store this guard in a variable that lives for the duration of your program. When it’s dropped, the trace file will be completed and closed.

Aliased Type

struct ChromeTraceGuard { /* private fields */ }