linera_service::util

Function wait_for_next_round

pub async fn wait_for_next_round(
    stream: &mut Pin<Box<dyn Stream<Item = Notification> + Send>>,
    timeout: RoundTimeout,
)

Returns after the specified time or if we receive a notification that a new round has started.