Crate linera_service

This module provides the executables needed to operate a Linera service, including a placeholder wallet acting as a GraphQL service for user interfaces.

Modules

cli_wrappers

Helper module to call the binaries of linera-service with appropriate command-line arguments.

faucet

node_service

project

prometheus_server

util

wallet

Macros

test_name