linera_service

Module cli_wrappers

Helper module to call the binaries of linera-service with appropriate command-line arguments.

Modules

docker

How to run Docker operations

local_kubernetes_net

How to run Linera validators locally as a Kubernetes deployment.

local_net

How to run Linera validators locally as native processes.

remote_net

How to connect to running GCP devnet.

Structs

ApplicationWrapper

A running Application to be queried in GraphQL.

ClientWrapper

Wrapper to run a Linera client command.

Faucet

A faucet instance that can be queried.

FaucetService

A running faucet service.

NetworkConfig

Network protocol in use outside and inside a Linera net.

NodeService

A running node service.

Enums

Network

Network protocol in use

OnClientDrop

Action to perform when the ClientWrapper is dropped.

Traits

LineraNet

A running Linera net.

LineraNetConfig

The information needed to start a Linera net of a particular kind.

NotificationsExt