linera_service::tracing::chrome

Function init

pub fn init(
    log_name: &str,
    writer: impl Write + Send + 'static,
) -> ChromeTraceGuard

Initializes tracing with Chrome Trace JSON exporter.

Returns a guard that must be kept alive for the duration of the program. When the guard is dropped, the trace data is flushed and completed.

Exports traces to Chrome Trace JSON format which can be visualized in:

Chrome: chrome://tracing

Perfetto UI: https://ui.perfetto.dev

Note: Uses try_init() to avoid panicking if a global subscriber is already set. In that case, tracing may not work as expected.